backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
anthropic
pdfplumber
python-docx
python-dotenv
slowapi
